# backend/openapi.yml
openapi: 3.1.0
info:
  title: RNA Analysis API
  version: 1.0.0
  description: API for RNA sequence analysis and retrieving results in SVG format.

paths:
  /api/analyze/pair:
    post:
      summary: Request for comparative analysis of wild type and mutant sequences
      description: Allows the user to request a comparative analysis of wild type and mutant RNA sequences.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                wildSequence:
                  type: string
                  description: Wild type RNA sequence.
                mutantSequence:
                  type: string
                  description: Mutant RNA sequence with SNP.
                analysis_id:
                  type: string
                  format: uuid
                  description: Unique analysis identifier.
              required:
                - wildSequence
                - mutantSequence
                - analysis_id
      responses:
        '200':
          description: Analysis request accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  analysis_id:
                    type: string
                    format: uuid
                    description: Unique analysis identifier.
        '400':
          description: Invalid input data.
        '500':
          description: Server error while processing the analysis request.

  /api/analyze/single:
    post:
      summary: Request for single sequence analysis
      description: Request analysis for a single RNA sequence to identify possible SNPs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                wildSequence:
                  type: string
                  description: RNA sequence for analysis.
              required:
                - wildSequence
      responses:
        '200':
          description: Analysis request accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  analysis_id:
                    type: string
                    format: uuid
                    description: Unique analysis identifier.
        '400':
          description: Invalid input data.
        '500':
          description: Server error while processing the analysis request.

  /api/results/{analysis_id}:
    get:
      summary: Retrieve analysis results
      description: Retrieve the analysis results for a given identifier.
      parameters:
        - name: analysis_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique analysis identifier.
      responses:
        '200':
          description: Analysis results retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResults'
        '404':
          description: Analysis results not found.
        '500':
          description: Server error.

components:
  schemas:
    AnalysisResults:
      type: object
      properties:
        analysis_id:
          type: string
          format: uuid
          description: Unique analysis identifier.
        results:
          type: object
          properties:
            RNApdist:
              type: number
              description: RNApdist result.
            RNAdistance:
              type: object
              properties:
                f:
                  type: number
                h:
                  type: number
                w:
                  type: number
                c:
                  type: number
                F:
                  type: number
                H:
                  type: number
                W:
                  type: number
                C:
                  type: number
                P:
                  type: number
            RNAfold:
              type: object
              properties:
                wt-dotbracket:
                  type: string
                mut-dotbracket:
                  type: string
            RNAplot:
              type: object
              properties:
                wt-svg:
                  type: string
                  description: Link to the wild type SVG.
                mut-svg:
                  type: string
                  description: Link to the mutant SVG.
            HIT_tree:
              type: object
              properties:
                tree:
                  type: string
                  description: Link to the HIT tree.
            SNP_variants:
              type: object
              properties:
                variantsfile:
                  type: string
                  description: Link to the SNP variants file.
            status:
              type: string
              description: Analysis status.